begin adding a general purpose universal software renderer. it doesnt do much yet, but osd can use it to draw its text more slowly.

+++ trunk/desmume/src/GPU_osd.h
@@ -32,7 +32,7 @@
 class OSDCLASS
 {
 private:
-	u8		screen[256*192*2];
+	u16		screen[256*192*2];
 	u64		offset;
 	u8		mode;
